solve sonar issue about code duplication
