+++ cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/elements/EquivalentBranchConversion.java
@@ -14,29 +14,10 @@ import com.powsybl.triplestore.api.PropertyBag;
 /**
  * @author Luma Zamarreño <zamarrenolm at aia.es>
  */
-public class EquivalentBranchConversion extends AbstractConductingEquipmentConversion {
+public class EquivalentBranchConversion extends AbstractBranchConversion {
 
     public EquivalentBranchConversion(PropertyBag b, Conversion.Context context) {
-        super("EquivalentBranch", b, context, 2);
-    }
-
-    @Override
-    public boolean valid() {
-        if (!super.valid()) {
-            return false;
-        }
-        String node1 = nodeId(1);
-        String node2 = nodeId(2);
-        if (context.boundary().containsNode(node1)
-                || context.boundary().containsNode(node2)) {
-            invalid("Has TopologicalNode on boundary");
-            return false;
-        }
-        if (!p.containsKey("r") || !p.containsKey("x")) {
-            invalid("No r,x attribures");
-            return false;
-        }
-        return true;
+        super("EquivalentBranch", b, context);
     }
 
     @Override
