+++ cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/elements/SeriesCompensatorConversion.java
@@ -15,29 +15,10 @@ import com.powsybl.triplestore.api.PropertyBag;
 /**
  * @author Luma Zamarreño <zamarrenolm at aia.es>
  */
-public class SeriesCompensatorConversion extends AbstractConductingEquipmentConversion {
+public class SeriesCompensatorConversion extends AbstractBranchConversion {
 
     public SeriesCompensatorConversion(PropertyBag sc, Conversion.Context context) {
-        super(CgmesNames.SERIES_COMPENSATOR, sc, context, 2);
-    }
-
-    @Override
-    public boolean valid() {
-        if (!super.valid()) {
-            return false;
-        }
-        String node1 = nodeId(1);
-        String node2 = nodeId(2);
-        if (context.boundary().containsNode(node1)
-                || context.boundary().containsNode(node2)) {
-            invalid("Has TopologicalNode on boundary");
-            return false;
-        }
-        if (!p.containsKey("r") || !p.containsKey("x")) {
-            invalid("No r,x attribures");
-            return false;
-        }
-        return true;
+        super(CgmesNames.SERIES_COMPENSATOR, sc, context);
     }
 
     @Override
